refa:refactor NotionApiService to its own package
